warnings
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@1585 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR0/HWSVMR0.cpp
@@ -513,7 +513,7 @@ HWACCMR0DECL(int) SVMR0LoadGuestState(PVM pVM, CPUMCTX *pCtx)
 HWACCMR0DECL(int) SVMR0RunGuestCode(PVM pVM, CPUMCTX *pCtx)
 {
     int         rc = VINF_SUCCESS;
-    uint64_t    exitCode = SVM_EXIT_INVALID;
+    uint64_t    exitCode = (uint64_t)SVM_EXIT_INVALID;
     SVM_VMCB   *pVMCB;
     bool        fForceTLBFlush = false;
 
@@ -651,7 +651,7 @@ ResumeExecution:
     /* Reason for the VM exit */
     exitCode = pVMCB->ctrl.u64ExitCode;
 
-    if (exitCode == SVM_EXIT_INVALID)          /* Invalid guest state. */
+    if (exitCode == (uint64_t)SVM_EXIT_INVALID)          /* Invalid guest state. */
     {
         HWACCMDumpRegs(pCtx);
 #ifdef DEBUG
